Correctly prefix path when searching for files
The search_file tool will search for a file using Dir.glob and include a
base path if there is one. It will then base the result into
read_contents. However, Dir.glob with a base will return results
relative to that base, so the read contents will fail.

Many times the LLM will see the resulting error and then use the full
path with a base of ., which will work correctly. This should just skip
that roundtrip.