Add color to minitest
I find it super hard to easily skim tests unless they have color. Every
more so when there is a lot of noisy output like the test suite currently
has.

This adds minitest-rg, which is an official minitest extension. It will
turn on color output when it's on a TTY by default. We also can turn it
on during CI because GitHub Actions supports it.