Remove workaround for test order
I was seeing this warning during tests:

    activesupport-8.0.2/lib/active_support.rb:100: warning: previous definition of test_order was here

I checked the git blame, and don't see any background on why this is
needed. I'm not sure if it's a bug in ActiveSupport or a bug in our
test suite.

This tries to remove it to see if it's still a problem. Some cursory
running of the tests does seem to be random (ie different seeds) at
least.