Copy example workflows in test sandboxes (#386)
Adds an easy way to import workflow examples from the project `examples/` directory when using the `in_sandbox` test function. If the symbol passed to `with_workflow` matches a directory in `examples/`, the entire example will be copied into the `roast` directory in the sandbox.

This allows us to validate our examples via testing. If a feature would break an example, tests will fail.

Additionally, testing the examples by definition is testing features we care about and expose to users, in the way we expect them to be used.

Examples can be tested as follows:

```ruby
  test "example LLM workflow" do
    in_sandbox with_workflow: :apply_diff_demo do
      roast("execute", "apply_diff_demo")
    end
  end
```